remove unsed components, created routes between pages, styled home page, added validation for City input
