+++ package.json
@@ -22,10 +22,9 @@
     "dayjs": "^1.11.10",
     "lottie-react": "^2.4.0",
     "react": "^18.2.0",
-    "react-bootstrap-date-picker": "^5.1.0",
     "react-datepicker": "^6.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.22.1"
+    "react-router-dom": "^6.22.2"
   },
   "devDependencies": {
     "@types/react": "^18.2.56",
