+++ src/App.jsx
@@ -1,16 +1,11 @@
+import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Header from "./components/Header";
-import Hero from "./components/Hero";
 import Footer from "./components/Footer";
 import Home from "./pages/Home";
-import Results from "./pages/Results";
-import Event from "./components/Event";
-import Contactform from "./components/Contactform";
+import About from "./pages/About";
+import Contact from "./pages/Contact";
 import Navbar from "./components/Navbar";
-// import Searchbar from "./components/Searchbar";
-import Eventcard from "./components/Eventcard";
-import Upcomingevents from "./components/Upcomingevents";
-import About from "./components/About";
 import "./App.css";
 
 function App() {
@@ -20,21 +15,12 @@ function App() {
         <Header>
           <Navbar />
         </Header>
-        <Hero />
-        {/* <Searchbar /> */}
-        {/* <Eventcard /> */}
-
-        <Upcomingevents />
-        {/* <About /> */}
-        {/* <Contactform /> */}
         <Routes>
-          <Route path="/results" element={<Results />} />
-          <Route path="/event" element={<Event />} />
-          {/* <Route path="/contact" element={<Contact />} /> */}
-          <Route path="*" element={<Home />} />
+          <Route path="/" element={<Home />} />
+          <Route path="/about" element={<About />} />
+          <Route path="/contact" element={<Contact />} />
         </Routes>
       </Router>
-
       <Footer />
     </>
   );
