--- src/components/About/index.jsx
@@ -1,23 +1,0 @@
-import React from "react";
-import HeroMini from "../HeroMini";
-
-const About = () => {
-  return (
-    <>
-    < HeroMini />
-      <h2>About Us</h2>
-      <br />
-      <p>
-        Lorem ipsum dolor sit amet consectetur adipisicing elit. Delectus sed
-        nisi obcaecati nam veritatis eum! Sed provident libero suscipit non
-        quibusdam debitis nobis, iste doloremque quos mollitia quas minus amet?
-        Lorem ipsum dolor sit amet consectetur adipisicing elit. Excepturi
-        dolorum itaque dignissimos assumenda iusto ratione, reprehenderit
-        corporis delectus praesentium, velit repellat totam, nostrum minima?
-        Totam enim omnis quaerat sapiente laborum? Lorem, ipsum dolor.
-      </p>
-    </>
-  );
-};
-
-export default About;
