--- src/components/Eventcard/eventcard.css
@@ -1,46 +1,0 @@
-.no-events {
-  background: #fff;
-  min-height: 110px;
-  border-radius: 4px;
-}
-
-.no-events p {
-  color: #f44336;
-  font-weight: 500;
-  font-size: 20px;
-}
-
-.kids {
-  background: #64312d;
-}
-
-.btn-event {
-  padding: 8px 14px;
-  border-radius: 5px;
-  border: 1px solid #f44336;
-  background: #f44336;
-  color: #fff;
-  text-transform: uppercase;
-  font-weight: 500;
-}
-
-.price-from,
-.price-to {
-  display: flex;
-  align-items: center; /* Vertically center the content */
-  justify-content: center; /* Horizontally center the content */
-  height: 100%; /* Set the height of the container */
-  margin: 0; /* Remove any default margin */
-}
-
-.event-note-text {
-  font-size: 12px;
-}
-.event-note-text i {
-  color: #f44336;
-}
-@media (max-width: 576px) {
-  .mui-header-title {
-    font-size: 16px;
-  }
-}
