+++ src/components/Footer/footer.css
@@ -1,7 +1,7 @@
 :root {
-  --primary-color: #9BB8CD;
+  --primary-color: #9bb8cd;
   --secondary-color: #ede8d2;
-  --third-color: #EEC759;
+  --third-color: #eec759;
   --link-hover: #ecf1e1c8;
   --back-drop: #ffffffc8;
   --dark: #17283d;
