--- src/components/HeroMini/HeroMini.css
@@ -1,11 +1,0 @@
-#jumbo-mini {
-    background-image: url("/public/artsyEvent.jpg");
-    background-size: cover;
-    background-position: center;
-    background-repeat: no-repeat;
-}
-
-#mini-overlay {
-    background-color: #00000080;
-    padding: 1rem;
-}
