+++ src/pages/About/index.jsx
@@ -1,0 +1,14 @@
+import Hero from "../../components/Hero";
+import "./about.css";
+import React, { useState } from "react";
+
+function About() {
+  return (
+    <>
+      <Hero />
+      <h1>About page</h1>
+    </>
+  );
+}
+
+export default About;
