+++ src/pages/Contact/index.jsx
@@ -1,0 +1,13 @@
+import "./contact.css";
+import React, { useState } from "react";
+import Contactform from "../../components/Contactform";
+function Contact() {
+  return (
+    <>
+      <Contactform />
+      <h1>Contact page</h1>
+    </>
+  );
+}
+
+export default Contact;
