--- src/pages/Results/index.jsx
@@ -1,9 +1,0 @@
-import "./results.css";
-
-import React from "react";
-
-const Results = () => {
-  return <div>Results</div>;
-};
-
-export default Results;
